converted iyk api to react query hook
